SRB2: fixed big endian builds
git-svn-id: svn://code.srb2.org/SRB2/trunk@4743 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/d_net.c
@@ -731,7 +731,7 @@ static void DebugPrintpacket(const char *header)
 			fprintf(debugfile, "    firsttic %u ply %d tics %d ntxtcmd %"PRIdS"\n    ",
 				(UINT32)ExpandTics(netbuffer->u.serverpak.starttic), netbuffer->u.serverpak.numslots,
 				netbuffer->u.serverpak.numtics,
-				(&((byte *)netbuffer)[doomcom->datalength] - (byte *)&netbuffer->u.serverpak.cmds[netbuffer->u.serverpak.numslots*netbuffer->u.serverpak.numtics]));
+				(size_t)(&((byte *)netbuffer)[doomcom->datalength] - (byte *)&netbuffer->u.serverpak.cmds[netbuffer->u.serverpak.numslots*netbuffer->u.serverpak.numtics]));
 			fprintfstring((char *)&netbuffer->u.serverpak.cmds[netbuffer->u.serverpak.numslots*netbuffer->u.serverpak.numtics],(byte)(
 				&((byte *)netbuffer)[doomcom->datalength] - (byte *)&netbuffer->u.serverpak.cmds[netbuffer->u.serverpak.numslots*netbuffer->u.serverpak.numtics]));
 			break;
